requirements.txt
Django==4.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.10
celery==5.5.3
redis==5.2.1
python-dotenv==1.1.1


